implemented dynamic pattern
