+++ src/mode.rs
@@ -1,25 +1,45 @@
 mod continuous;
+pub(crate) mod dynamic;
 mod powered_off;
 mod ready;
 
-use core::convert::TryFrom;
-
 pub use continuous::*;
+pub use dynamic::*;
 use embedded_hal::blocking::i2c::{Write, WriteRead};
 pub use powered_off::*;
 pub use ready::*;
 
 use crate::error::Error;
-use crate::register::{
-    AmbientStatusErrorCode, RangeStatusErrorCode, Register16Bit, Register8Bit,
-    ResultInterruptStatusGpioCode, SysAmbientStartCode, SysInterruptClearCode, SysRangeStartCode,
-};
 use crate::VL6180X;
+
+impl<MODE, I2C, E> VL6180X<MODE, I2C>
+where
+    I2C: WriteRead<Error = E> + Write<Error = E>,
+{
+    fn into_mode<MODE2>(self, mode: MODE2) -> VL6180X<MODE2, I2C> {
+        VL6180X {
+            mode,
+            com: self.com,
+            config: self.config,
+            state: self.state,
+        }
+    }
+}
+
+/// Operating modes with this trait have an implementation to initialize VL6180X into that mode.
+pub trait AllowInit {
+    /// Create a new instance of itself.
+    fn new() -> Self;
+}
+
 /// Operating modes with this trait have an implementation for reading measurements
 pub trait AllowReadMeasurement {}
+
 /// Operating modes with this trait have an implementation for starting a single
 /// ambient light measurement
+
 pub trait AllowStartAmbientSingle {}
+
 /// Operating modes with this trait have an implementation for starting a single
 /// range measurement
 pub trait AllowStartRangeSingle {}
@@ -32,95 +52,27 @@ where
     /// Blocking read of the range mesurement.
     /// The reading (whether single or continuous) must already have been started.
     pub fn read_range_blocking_mm(&mut self) -> Result<u8, Error<E>> {
-        // TODO: convert timeout to be in millis instead of loops.
-        let mut c = 0;
-        while !ResultInterruptStatusGpioCode::has_error_or_event(
-            ResultInterruptStatusGpioCode::NewSampleReadyRangeEvent,
-            self.read_named_register(Register8Bit::RESULT__INTERRUPT_STATUS_GPIO)?,
-        ) {
-            c += 1;
-            if c == self.config.io_timeout {
-                self.state.did_timeout = true;
-                return Err(Error::Timeout);
-            }
-        }
-
-        self.get_range_val_and_status()
+        self.read_range_blocking_mm_direct()
     }
 
     /// Non-blocking read of the range measurement.
     /// The reading (whether single or continuous) must already have been started.
     /// Returns [Error::ResultNotReady] if the result is not ready.
     pub fn read_range_mm(&mut self) -> Result<u8, Error<E>> {
-        if !ResultInterruptStatusGpioCode::has_error_or_event(
-            ResultInterruptStatusGpioCode::NewSampleReadyRangeEvent,
-            self.read_named_register(Register8Bit::RESULT__INTERRUPT_STATUS_GPIO)?,
-        ) {
-            return Err(Error::ResultNotReady);
-        }
-        self.get_range_val_and_status()
-    }
-
-    fn get_range_val_and_status(&mut self) -> Result<u8, Error<E>> {
-        let status = self.read_named_register(Register8Bit::RESULT__RANGE_STATUS)?;
-        let error = RangeStatusErrorCode::try_from(status)
-            .map_err(|_| Error::UnknownRegisterCode(status))?;
-        if error != RangeStatusErrorCode::NoError {
-            return Err(Error::RangeStatusError(error));
-        }
-        let range = self.read_named_register(Register8Bit::RESULT__RANGE_VAL)?;
-        self.write_named_register(
-            Register8Bit::SYSTEM__INTERRUPT_CLEAR,
-            SysInterruptClearCode::ClearRangeInterrupt as u8,
-        )?;
-        Ok(range)
+        self.read_range_mm_direct()
     }
 
     /// Blocking read of the range mesurement.
     /// The reading (whether single or continuous) must already have been started.
     pub fn read_ambient_blocking(&mut self) -> Result<u16, Error<E>> {
-        // TODO: convert timeout to be in millis instead of loops.
-        let mut c = 0;
-        while !ResultInterruptStatusGpioCode::has_error_or_event(
-            ResultInterruptStatusGpioCode::NewSampleReadyAmbientEvent,
-            self.read_named_register(Register8Bit::RESULT__INTERRUPT_STATUS_GPIO)?,
-        ) {
-            c += 1;
-            if c == self.config.io_timeout {
-                self.state.did_timeout = true;
-                return Err(Error::Timeout);
-            }
-        }
-        self.get_ambient_val_and_status()
+        self.read_ambient_blocking_direct()
     }
 
     /// Non-blocking read of the ambient light measurement.
     /// The reading (whether single or continuous) must already have been started.
     /// Returns [Error::ResultNotReady] if the result is not ready.
     pub fn read_ambient(&mut self) -> Result<u16, Error<E>> {
-        if !ResultInterruptStatusGpioCode::has_error_or_event(
-            ResultInterruptStatusGpioCode::NewSampleReadyAmbientEvent,
-            self.read_named_register(Register8Bit::RESULT__INTERRUPT_STATUS_GPIO)?,
-        ) {
-            return Err(Error::ResultNotReady);
-        }
-        self.get_ambient_val_and_status()
-    }
-
-    fn get_ambient_val_and_status(&mut self) -> Result<u16, Error<E>> {
-        let status = self.read_named_register(Register8Bit::RESULT__ALS_STATUS)?;
-        if status != AmbientStatusErrorCode::NoError as u8 {
-            let error = AmbientStatusErrorCode::try_from(status)
-                .map_err(|_| Error::UnknownRegisterCode(status))?;
-            return Err(Error::AmbientStatusError(error));
-        }
-        let ambient = self.read_named_register_16bit(Register16Bit::RESULT__ALS_VAL)?;
-        self.write_named_register(
-            Register8Bit::SYSTEM__INTERRUPT_CLEAR,
-            SysInterruptClearCode::ClearAmbientInterrupt as u8,
-        )?;
-
-        Ok(ambient)
+        self.read_ambient_direct()
     }
 }
 
@@ -138,10 +90,7 @@ where
     /// 3. Wait for the ambient interrupt to be triggered, indicating that the
     /// new sample is ready, then call [`read_ambient`](#method.read_ambient).
     pub fn start_ambient_single(&mut self) -> Result<(), E> {
-        self.write_named_register(
-            Register8Bit::SYSALS__START,
-            SysAmbientStartCode::SingleStart as u8,
-        )
+        self.start_ambient_single_direct()
     }
 }
 
@@ -159,9 +108,6 @@ where
     /// 3. Wait for the range interrupt to be triggered, indicating that the
     /// new sample is ready, then call [`read_range`](#method.read_range).
     pub fn start_range_single(&mut self) -> Result<(), E> {
-        self.write_named_register(
-            Register8Bit::SYSRANGE__START,
-            SysRangeStartCode::SingleStart as u8,
-        )
+        self.start_range_single_direct()
     }
 }
