+++ src/mode/ready.rs
@@ -1,16 +1,21 @@
 use crate::error::Error;
-use crate::register::{Register8Bit, SysAmbientStartCode, SysRangeStartCode};
 use crate::VL6180X;
 use embedded_hal::blocking::i2c::{Write, WriteRead};
 
 use super::{
-    AllowReadMeasurement, AllowStartAmbientSingle, AllowStartRangeSingle, AmbientContinuousMode,
-    InterleavedContinuousMode, RangeContinuousMode,
+    AllowInit, AllowReadMeasurement, AllowStartAmbientSingle, AllowStartRangeSingle,
+    AmbientContinuousMode, InterleavedContinuousMode, RangeContinuousMode,
 };
 /// Sensor has been configured and is ready to take single measurements or switch to a
 /// continuous measurement mode
 #[derive(Debug, Copy, Clone)]
 pub struct ReadyMode;
+
+impl AllowInit for ReadyMode {
+    fn new() -> Self {
+        Self {}
+    }
+}
 
 impl AllowReadMeasurement for ReadyMode {}
 
@@ -26,22 +31,14 @@ where
     /// Starts a single range measurement then calls [`read_range_blocking`](#method.read_range_blocking)
     /// to wait for the result.
     pub fn poll_range_single_blocking_mm(&mut self) -> Result<u8, Error<E>> {
-        self.write_named_register(
-            Register8Bit::SYSRANGE__START,
-            SysRangeStartCode::SingleStart as u8,
-        )?;
-        self.read_range_blocking_mm()
+        self.poll_range_single_blocking_mm_direct()
     }
 
     /// Poll the sensor for a single ambient light measurement.
     /// Starts a single ambient measurement then calls [`read_ambient_blocking`](#method.read_ambient_blocking)
     /// to wait for the result.
     pub fn poll_ambient_single_blocking(&mut self) -> Result<u16, Error<E>> {
-        self.write_named_register(
-            Register8Bit::SYSALS__START,
-            SysAmbientStartCode::SingleStart as u8,
-        )?;
-        self.read_ambient_blocking()
+        self.poll_ambient_single_blocking_direct()
     }
 
     /// Starts continuous operation mode for reading range measurements.
@@ -53,7 +50,7 @@ where
         self,
     ) -> Result<VL6180X<RangeContinuousMode, I2C>, Error<E>> {
         let mut new_vl6180x = self.into_mode(RangeContinuousMode {});
-        new_vl6180x.toggle_range_continuous()?;
+        new_vl6180x.toggle_range_continuous_direct()?;
         Ok(new_vl6180x)
     }
 
@@ -62,7 +59,7 @@ where
         self,
     ) -> Result<VL6180X<AmbientContinuousMode, I2C>, Error<E>> {
         let mut new_vl6180x = self.into_mode(AmbientContinuousMode {});
-        new_vl6180x.toggle_ambient_continuous()?;
+        new_vl6180x.toggle_ambient_continuous_direct()?;
         Ok(new_vl6180x)
     }
 
@@ -72,7 +69,7 @@ where
         self,
     ) -> Result<VL6180X<InterleavedContinuousMode, I2C>, Error<E>> {
         let mut new_vl6180x = self.into_mode(InterleavedContinuousMode {});
-        new_vl6180x.enable_interleaved_continuous()?;
+        new_vl6180x.enable_interleaved_continuous_direct()?;
         Ok(new_vl6180x)
     }
 }
