+++ src/register.rs
@@ -245,3 +245,8 @@ impl TryFrom<u8> for AmbientStatusErrorCode {
         AmbientStatusErrorCode::from_int(code >> 4)
     }
 }
+// RANGE_SCALER values for 1x, 2x, 3x scaling - see STSW-IMG003 core/src/vl6180x_api.c (ScalerLookUP[])
+pub const RANGE_SCALAR_CODE: [u16; 4] = [0, 253, 127, 84];
+/// See datasheet 2.10.6 for more details
+pub const AMBIENT_ANALOGUE_GAIN_CODE: [u8; 8] = [0x46, 0x45, 0x44, 0x43, 0x42, 0x41, 0x40, 0x47];
+pub const AMBIENT_ANALOGUE_GAIN_VALUE: [f32; 8] = [1.01, 1.28, 1.72, 2.60, 5.21, 10.32, 20.0, 40.0];
